refactor: Day-2 move logic to standalone function
The executeSecondHalf was really meatier than it should have
been, since there were iterations involved. Refactored to pull
out the function to a standalone function.
