+++ src/day-2/intcode.ts
@@ -1,3 +1,8 @@
+export interface ProgramMatch {
+  noun: number;
+  verb: number;
+}
+
 export function runProgram(commands: number[]): number[] {
   for (let i = 0; i < commands.length; i += 4) {
     const command = commands[i];
@@ -7,11 +12,36 @@ export function runProgram(commands: number[]): number[] {
     if (command === 99) {
       break;
     }
-    if (command === 1) {
-      commands[resultIndex] = commands[firstIndex] + commands[secondIndex];
-    } else if (command === 2) {
-      commands[resultIndex] = commands[firstIndex] * commands[secondIndex];
-    }
+
+    commands[resultIndex] =
+      command === 1
+        ? commands[firstIndex] + commands[secondIndex]
+        : commands[firstIndex] * commands[secondIndex];
   }
   return commands;
+}
+
+export function findMatchingInput(
+  desiredResult: number,
+  commands: number[]
+): ProgramMatch {
+  let noun = 0;
+  let verb = 0;
+  let completed = false;
+  for (noun = 0; noun < 100; noun++) {
+    for (verb = 0; verb < 100; verb++) {
+      const memory = commands.slice();
+      memory[1] = noun;
+      memory[2] = verb;
+      const result = runProgram(memory)[0];
+      if (result === desiredResult) {
+        completed = true;
+        break;
+      }
+    }
+    if (completed) {
+      break;
+    }
+  }
+  return { noun, verb };
 }
