+++ tslint.json
@@ -4,6 +4,7 @@
     "tslint-config-prettier"
   ],
   "rules": {
-    "object-literal-sort-keys": false
+    "object-literal-sort-keys": false,
+    "interface-name": false
   }
 }
